Add first version of a standard test runner
When finished this test runner will only require that each transit
implementation have a script in bin/roundtrip which can read a transit
string from standard in, roundtrip it, and write the result to standard
out.

It will feed all of the sample files as well as generated edn data.